// frontend/src/api/generated/schemas/dataTableStatus.ts
// @ts-nocheck
// This file is auto-generated by orval. Do not edit manually.
import type { DataTableStatusDateRange } from "./dataTableStatusDateRange";

/**
 * Status of a data table.
 */
export interface DataTableStatus {
  date_range?: DataTableStatusDateRange;
  name: string;
  record_count: number;
  status?: string;
}